+++ src/main/java/com/deiz0n/junit/services/GenericService.java
@@ -7,10 +7,10 @@ import java.util.List;
 
 public interface GenericService {
 
-    User getResource(Long id);
+    User getResource(Integer id);
     List<User> getResources();
     User createResource(UserDTO newUser);
     User updateResource(UserDTO newUser);
-    void removeResource(Long id);
+    void removeResource(Integer id);
 
 }
